ui: Add autocomplete feature
Autocomplete is now supported, starting to type a path in the
inputbox (ChangeDirInput) will open a dropdown list where a directory
can be selected to change directory to. Symlinked directories, however,
will not be listed, but when typing its path manually, autocomplete
works. This will be fixed in future commits