ui: Improve autocomplete
Autocomplete is more responsive with this commit, and the
keybindings were modified to make it easier to control/edit the
autocompletion input.

In directories with a huge amount of entries, however, autocompletion
will lag. This happens only on the device side (i.e ADB mode), where
there is significant latency in transferring and processing the directory
listing to the client. This will be fixed in future commits.