ui: Set selector position correctly after any operation
Prior to this commit, after any operation (especially move/delete)
the selector would always be placed at the first row, making it tedious
to come back to the previous entry position in the current directory.
Launching ChangeDirInput also had the same effect, since pressing 'g'
triggers tview's default behavior of scrolling to the first entry.

This commit fixes this behavior by checking if the total number of
entries has been reduced, and places the selector appropriately, and
changing the default behavior of the 'g' key.