ui: Avoid locking a pane when operation is pending
Locking a pane after selecting entries can sometimes be a little too
restricting, especially if the user wants to open multiple folders and
select entries, or wants to select entries from both the device and the
client side and copy them to a destination, for example.

This commit introduces a new and better way of selecting and starting
operations, and in a way emulates nnn's behavior for the above-mentioned
purpose. Since multiselection was already implemented, all the user has
to do is to select as many files as they wish, no matter the mode or
pane, and press the appropriate keys to start the desired operation.