local, adb: Handle symlinked directories correctly
With this commit, symlinks are now properly handled, both on local
and adb modes.

However, symlinked directories are properly listed only in local mode.
In adb mode, trying to query whether a symlink is a directory or not is
currently done by running commands directly on the device, which results
in a very slow directory listing on the client, so entries which are
symlinks are shown as-is for now. They can, however, be accessed if selected.