local_utils: Ensure filtered entries are properly checked
Previously, trying to change directory after selected a filtered entry,
would produce unexpected behavior, such as attempting to update pathList
(change directory) even when the selected entry was a file, and was
because of trying to validate an entry using the wrong index.

This commit now maps the cells directly to pathList's data, so that
entries can be properly vetted before trying to change directories.