modal: Remove external dependency for fuzzy searching
The fuzzy package causes a lot of spikes in CPU usage, for small and
large lists of entries alike. Switching to the conventional way of
entry searching lead to a dramatic decrease in CPU usage, and also uses
less memory now, since we are directly querying pathList in showFilterInput,
and storing selected data in just one map instead of a separate slice and map
in showEditSelections.

Granted, searching will be less efficient, but in this case, the tradeoff is worth it.