local_utils: Make entry checking more efficient
In situations where we go to an empty directory, if there are no values,
a stale entry will be checked, and incorrectly processed. Avoid this by
making the cached entry nil if no referenced data has been found, and add
checks for this nil entry to ensure that only the correct data gets processed.