adb_utils: Make in-device copy operations cancellable
The device.RunCommand method does not allow for cancellation, hence,
whenever a copy operation was started in the device, it wouldn't be
cancellable unless the user killed the entire application.

Since all the checks for the client and device are already done in
getAdb(), we can directly invoke adb and provide a cancellable context
via exec.CommandContext.