ui: Implement opening files externally
With this commit, files of any format located on the client or the
device can be opened on the client side. The current method to open
files is via xdg-open, a config option will be provided in later commits.

Note that in some cases, applications may take time to exit, so the
opened file will not be released until the application exits, and any
attempts to perform operations on the file will return an error.

For example, after opening a zip file via file-roller, modifying it and
closing the file-roller GUI, file-roller takes some time to fully exit,
and the since the UI is waiting for file-roller to exit, the user cannot
perform operations on the currently modified file until file-roller exits.