ui: Implement custom list for ChangeDirInput
The list provided by tview's autocomplete is a bit buggy. While listing
large directories the list overlaps with the text area for instance, and
the user is unable to view the inputfield's text. Also the list's height
can vary with respect to the number of entries, and we currently need to
avoid this particular behavior.

This commit introduces a custom resizable (by width) list with a fixed
height, along with custom autocompletion and filtering functionalities.
This should replace the autocompletion provided by tview and also give
the UI a better aesthetic.