ops: Do not clear opsView prior to cancelling all operations
Prior to this commit, the "Cancel all operations" functionality did
not work properly. This is because opsView was cleared and only then
was cancelAllOps() called, which wouldn't do anything since the context
for each operation was attached to their respective selectable TableCell,
and clearing opsView beforehand would "throw away" said contexts, thus
effectively not cancelling all the operations. This commit fixes this issue.

Some minor improvements were made, which are:
- Data races related to accessing opsView were fixed
- Iterating through the cells in opsView were made much faster,
  by skipping cells that are not selectable.